refactor: unify instance_id and channel_id types to u32
- Standardize all instance_id and channel_id to u32 across the codebase
- Update keyspace functions to use u32 parameters
- Update SQLx queries to decode directly to u32 (avoiding i64 → u32 casts)
- Update DashMap generic key types to u32
- Update Axum Path extractors and handlers to use u32
- Update DTO structs and API models to use u32
- Fix test fixtures to use numeric instance IDs instead of strings
- Remove unnecessary `as u32` type conversions throughout

This change aligns with OPC UA standards (UInt32 for NodeId) and
simplifies the codebase by using a consistent type for all identifiers.

+++ libs/voltage-comlink/src/traits.rs
@@ -268,7 +268,7 @@ pub enum ChannelCommand {
 /// Batch telemetry data for channel transmission
 #[derive(Debug, Clone)]
 pub struct TelemetryBatch {
-    pub channel_id: u16,
+    pub channel_id: u32,
     pub telemetry: Vec<(u32, f64, i64)>, // (point_id, raw_value, timestamp)
     pub signal: Vec<(u32, f64, i64)>,    // (point_id, raw_value, timestamp)
 }
@@ -283,7 +283,7 @@ pub struct ExtendedPointData {
     pub unit: String,
     pub description: String,
     pub telemetry_type: Option<FourRemote>,
-    pub channel_id: Option<u16>,
+    pub channel_id: Option<u32>,
 }
 
 /// Test channel parameters
@@ -322,7 +322,7 @@ pub trait ComBase: Send + Sync {
     fn name(&self) -> &str;
 
     /// Get channel ID
-    fn get_channel_id(&self) -> u16;
+    fn get_channel_id(&self) -> u32;
 
     /// Get channel status
     async fn get_status(&self) -> ChannelStatus;
